// src/api/generated/model/satisfactionScale.ts
/**
 * Generated by orval v7.17.2 🍺
 * Do not edit manually.
 * PurposePunch.Api
 * OpenAPI spec version: 1.0
 */

export type SatisfactionScale =
  (typeof SatisfactionScale)[keyof typeof SatisfactionScale];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const SatisfactionScale = {
  VeryDissatisfied: "VeryDissatisfied",
  SlightlyDissatisfied: "SlightlyDissatisfied",
  NeitherSatisfiedNorDissatisfied: "NeitherSatisfiedNorDissatisfied",
  SlightlySatisfied: "SlightlySatisfied",
  VerySatisfied: "VerySatisfied",
} as const;